429: Added integration test for ASHE earnings
