+++ build.sbt
@@ -21,8 +21,8 @@ libraryDependencies ++= Seq(
   javaWs,
   javaJpa,
   "org.eclipse.persistence" % "eclipselink" % "2.6.2",
-   "com.github.ONSdigital" % "dp-dd-backend-model" % "1.0.32",
-//  "uk.co.onsdigital.discovery" % "dd-model" % "1.0.32",
+   "com.github.ONSdigital" % "dp-dd-backend-model" % "1.0.33",
+//  "uk.co.onsdigital.discovery" % "dd-model" % "1.0.33",
   "org.postgresql" % "postgresql" % "9.4.1208.jre7",
   "org.apache.kafka" % "kafka-clients" % "0.10.1.0",
   "org.flywaydb" % "flyway-core" % "4.0.3",
